Relocate config for access and Clean Up before build
