+++ runtime/src/main/java/io/quarkiverse/jsonrpc/runtime/config/JsonRPCOpenRpcConfig.java
@@ -1,4 +1,6 @@
-package io.quarkiverse.jsonrpc.deployment.config;
+package io.quarkiverse.jsonrpc.runtime.config;
+
+import java.nio.file.Path;
 
 import io.quarkus.runtime.annotations.ConfigGroup;
 import io.quarkus.runtime.annotations.ConfigItem;
@@ -12,20 +14,32 @@ public class JsonRPCOpenRpcConfig {
     public boolean enabled;
 
     /**
+     * HTTP Path for the JsonRPC OpenRPC DevUI
+     */
+    @ConfigItem(defaultValue = "open-rpc")
+    public String basePath;
+
+    /**
      * HTTP Path for the JsonRPC OpenRPC Playground
      */
-    @ConfigItem(defaultValue = "/open-rpc/playground")
+    @ConfigItem(defaultValue = "playground")
     public String playgroundPath;
 
     /**
      * HTTP Path for the JsonRPC OpenRPC Spec file
      */
-    @ConfigItem(defaultValue = "/open-rpc/openrpc.json")
-    public String specPath;
+    @ConfigItem(defaultValue = "openrpc.json")
+    public String schemaPath;
 
     /**
      * If the OpenRPC Schema file should be accessible
      */
     @ConfigItem(defaultValue = "true")
     public boolean schemaAvailable;
+
+    /**
+     * Location to store the generated OpenRPC Schema file
+     */
+    @ConfigItem(defaultValue = "/open-rpc")
+    public Path storeSchemaDirectory;
 }
