+++ runtime/src/main/java/io/quarkiverse/jsonrpc/runtime/config/JsonRpcConfig.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.jsonrpc.deployment.config;
+package io.quarkiverse.jsonrpc.runtime.config;
 
 import io.quarkus.runtime.annotations.ConfigPhase;
 import io.quarkus.runtime.annotations.ConfigRoot;
@@ -7,7 +7,7 @@ import io.smallrye.config.WithName;
 /**
  * JsonRPC Configuration
  */
-@ConfigRoot(phase = ConfigPhase.BUILD_TIME)
+@ConfigRoot(name = "json-rpc", phase = ConfigPhase.BUILD_AND_RUN_TIME_FIXED)
 public class JsonRpcConfig {
 
     /**
